requirements.txt
evtx
netaddr
numpy
pandas
python-dateutil
pytz
six
XlsxWriter
flatten_json